@EditFeature: Made major changes in models. Separated activity from inventory. separated activity dates from activity. Basically made models InventoryActivity to store activities for a given inventory. For each Activity, a separate model called InventoryActivityAssignment( was made already ) was modified to store dates for each activity. For each date images are stored in InventoryActivityImage model. This was done because earlier dates were stored itself in inventory model which made difficult to query the dates per activity. Also audit dates were stoted in Audit Date table which kind of looked unpleasant.  Also since we have to store the assigned user against each activity and for each date, it makes sense to separate activity dates from activity itself.
The entries to InventoryActivity is made whenever user runs ConvertToCampaign API. For each inventory, three entries are made in InventoryActivity
table. For Release and Closure, date entry is made in InventoryActivityAssignment table. But not for Audit because it contains more than one
date for a given inventory in a general case.

The entries of InventoryActivityAssignment are deleted whenever user converts a campaign back  to proposal.

The GET api to fetch all the suppliers and inventories is modified to reflect the changes.
The PUT api to update dates, phase, status, comment, etc is modified to reflect the changes.